durations precision increased to 32 bits. this allows for precise delta frame times to be accepted without accumulation.

+++ include/tweenpoint.tcc
@@ -102,12 +102,12 @@ namespace tweeny {
         }
 
         template<typename... Ts>
-        inline uint16_t tweenpoint<Ts...>::duration() const {
+        inline uint32_t tweenpoint<Ts...>::duration() const {
             return *std::max_element(durations.begin(), durations.end());
         }
 
         template<typename... Ts>
-        inline uint16_t tweenpoint<Ts...>::duration(size_t i) const {
+        inline uint32_t tweenpoint<Ts...>::duration(size_t i) const {
             return durations.at(i);
         }
     }
